Refactor Client Commands into it's package.
The client commands are not in a package, but rather
in the vague directory of commands.

As we are adding more commands, this could mix up
with the other classes, causing fellow developers
confusion.

Moved the client commands into it's own package, and
edited the references according of different files
that depend on the client commands.

This will ensure that the directory is up to date
and as specific as possible for OOP reasons.
