+++ src/main/java/seedu/address/logic/commands/clients/ClearCommand.java
@@ -1,7 +1,8 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.clients;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.logic.commands.Command;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
