+++ src/main/java/seedu/address/logic/commands/clients/CommandResult.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.clients;
 
 import static java.util.Objects.requireNonNull;
