+++ src/main/java/seedu/address/logic/commands/clients/FindCommand.java
@@ -1,9 +1,10 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.clients;
 
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
+import seedu.address.logic.commands.Command;
 import seedu.address.model.Model;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
