+++ src/main/java/seedu/address/logic/commands/clients/ListCommand.java
@@ -1,8 +1,9 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.clients;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
+import seedu.address.logic.commands.Command;
 import seedu.address.model.Model;
 
 /**
