+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -8,15 +8,15 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListCommand;
+import seedu.address.logic.commands.clients.AddCommand;
+import seedu.address.logic.commands.clients.ClearCommand;
+import seedu.address.logic.commands.clients.DeleteCommand;
+import seedu.address.logic.commands.clients.EditCommand;
+import seedu.address.logic.commands.clients.ExitCommand;
+import seedu.address.logic.commands.clients.FindCommand;
+import seedu.address.logic.commands.clients.HelpCommand;
+import seedu.address.logic.commands.clients.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
