+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -17,6 +17,8 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.Messages;
+import seedu.address.logic.commands.clients.AddCommand;
+import seedu.address.logic.commands.clients.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
