+++ src/test/java/seedu/address/logic/commands/CommandResultTest.java
@@ -7,6 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.clients.CommandResult;
+
 public class CommandResultTest {
     @Test
     public void equals() {
