+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -13,15 +13,15 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
-import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListCommand;
+import seedu.address.logic.commands.clients.AddCommand;
+import seedu.address.logic.commands.clients.ClearCommand;
+import seedu.address.logic.commands.clients.DeleteCommand;
+import seedu.address.logic.commands.clients.EditCommand;
+import seedu.address.logic.commands.clients.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.commands.clients.ExitCommand;
+import seedu.address.logic.commands.clients.FindCommand;
+import seedu.address.logic.commands.clients.HelpCommand;
+import seedu.address.logic.commands.clients.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 import seedu.address.model.person.Person;
